change vars type

+++ include/Error.h
@@ -1,8 +1,6 @@
 #ifndef ERROR_H_
 #define ERROR_H_
 
-#include <Arduino.h>
-
 #define ERROR_MESSAGE_SIZE 90
 
 enum ErrorCode
@@ -21,7 +19,7 @@ struct Error
 {
     char errorMessage[ERROR_MESSAGE_SIZE];
     ErrorCode errorCode;
-    uint8_t printfError;
+    short printfError;
 };
 
 void addError(ErrorCode code, const char *fmt, ...);
